requirements.txt
eth-ape>=0.2.7
pytest>=6.2.5,<7.0.0
black>=21.5
requests==2.27.1
Flask==2.0.3
htmlmin==0.1.12
click>=8.1.3
pydantic>=1.9.1
pytest>=6.2.5
